+++ pkg/sso/sso.go
@@ -5,16 +5,47 @@ import (
 	"context"
 	"net/http"
 	"net/url"
+	"time"
+	"encoding/json"
+	
 )
 
-var (
-	DefaultClient *Client
-)
+// Client represents a client that performs Admin Portal requests to the WorkOS API.
+type Client struct {
+	// The WorkOS API Key. It can be found in https://dashboard.workos.com/api-keys.
+	APIKey string
 
-func Configure(apiKey, clientID string) {
-	DefaultClient.APIKey = apiKey
-	DefaultClient.ClientID = clientID
+	ClientID string
+
+	// The http.Client that is used to manage Admin Portal records from WorkOS.
+	// Defaults to http.Client.
+	HTTPClient *http.Client
+
+	// The endpoint to WorkOS API. Defaults to https://api.workos.com.
+	Endpoint string
+
+	// The function used to encode in JSON. Defaults to json.Marshal.
+	JSONEncode func(v interface{}) ([]byte, error)
 }
+
+// DefaultClient is the client used by SetAPIKey and mfa functions.
+var (
+	DefaultClient = NewClient("", "")
+) //question about this vs var DefaultClient *Client
+
+
+
+// NewClient returns a new instance of the Client struct with default values.
+func NewClient(apiKey, clientID string) *Client {
+    return &Client{
+        APIKey:     apiKey,
+        ClientID:   clientID,
+        Endpoint:   "https://api.workos.com",
+        HTTPClient: &http.Client{Timeout: time.Second * 10},
+        JSONEncode: json.Marshal,
+    }
+}
+
 
 // GetAuthorizationURL returns an authorization url generated with the given
 // options.
